Support building index funds via the vault
It allows admin fine grained control on how much to remove from each strategy in a withdrawal scenario. 
If the contract is used for building an index fund, the admin can use this feature to keep the token weightings inside the fund intact even after a withdrawal 
If strategyWithdrawalPercent = 1E18 for all strategies then the current withdrawal strategy of emptying one strategy before going to the next will be followed.